requirements.txt
pandas>=2.0.0
numpy>=1.24.0
scipy>=1.10.0
statsmodels>=0.14.0
scikit-learn>=1.3.0
networkx>=3.0
python-igraph>=0.11.0
leidenalg>=0.10.0
matplotlib>=3.7.0
seaborn>=0.12.0
adjustText>=1.3.0
